+++ meta_package_manager/__init__.py
@@ -23,6 +23,9 @@ import logging
 
 from .bitbar import fix_environment
 
+# Canonical name of the CLI.
+CLI_NAME = "mpm"
+
 __version__ = "3.3.1"
 """ Examples of valid version strings according :pep:`440#version-scheme`: